simple tamplate in header
